% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesFbar.R
\name{agesFbar}
\alias{agesFbar}
\title{Age Range for Fbar}
\usage{
agesFbar(cn, probs = c(0.1, 0.9), plot = FALSE, main = NULL)
}
\arguments{
\item{cn}{a vector, matrix, table, or data frame containing the catch at age.
Ages are indicated in names if \code{cn} is a vector, or in column
names if \code{cn} is a matrix, table, or data frame.}

\item{probs}{a vector of probabilities, e.g. \code{c(0.1, 0.9)} to calculate
the age range that covers 80\% of the catch.}

\item{plot}{whether to plot the cumulative distribution and interpolation.}

\item{main}{passed to \code{plot}.}
}
\value{
Vector of two values, the lower and upper age.
}
\description{
Calculate a suitable age range for average fishing mortality (Fbar).
}
\note{
\code{NA} values in \code{cn} are treated as zero.

The method uses linear interpolation of the cumulative distribution to
calculate the cut-off ages. See \code{plot = TRUE} example below.
}
\examples{
agesFbar(gss)
agesFbar(gss, plot=TRUE)
agesFbar(gss, plot=TRUE, main="Greater silver smelt in 5b and 6a")

# 50\% coverage instead of 80\%
agesFbar(gss, c(0.25, 0.75))

# Include only the last 20 years
agesFbar(tail(gss, 20))

}
\seealso{
\code{\link{gss}} is an example catch-at-age table.

\code{\link{cumsum}} calculates an empirical cumulative distribution.

\code{\link{approx}} performs linear interpolation.

\code{\link{icesAdvice-package}} gives an overview of the package.
}
