% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaCI.R
\name{sigmaCI}
\alias{sigmaCI}
\title{Sigma from Confidence Interval}
\usage{
sigmaCI(lo, hi, log = TRUE, level = 0.95)
}
\arguments{
\item{lo}{the lower confidence bound.}

\item{hi}{the upper confidence bound.}

\item{log}{whether the confidence interval is lognormal.}

\item{level}{the confidence level.}
}
\value{
Implicit value of sigma.
}
\description{
Calculate the implicit sigma that was used to construct a confidence
interval.
}
\note{
The purpose of PA reference points is to apply a precautionary approach in
fisheries management.

This function is useful for reviewing PA reference points, when the report
provides a CI but not the value of sigma.
}
\examples{
sigmaCI(100, 200)

}
\seealso{
\code{\link{sigmaPA}} calculates the implicit sigma from PA reference points.

\code{\link{icesAdvice-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
