% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesAdvice-package.R
\docType{package}
\name{icesAdvice-package}
\alias{icesAdvice}
\alias{icesAdvice-package}
\title{Functions Related to ICES Advice}
\description{
Provides functions that are related to the advisory work of ICES
(International Council for the Exploration of the Sea).
}
\details{
\emph{Calculate sigma:}
 \tabular{ll}{
   \code{\link{sigmaCI}} \tab from confidence interval\cr
   \code{\link{sigmaPA}} \tab from PA reference points
 }
}
\author{
Arni Magnusson, with contributions from Colin Millar.
}
\references{
ICES list of software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.

Overview of the ICES advisory process:
\url{http://ices.dk/community/advisory-process/Pages/default.aspx}.
}

