% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesAdvice-package.R
\docType{package}
\name{icesAdvice-package}
\alias{icesAdvice}
\alias{icesAdvice-package}
\title{Functions Related to ICES Advice}
\description{
Functions that are related to the advisory work of ICES (International
Council for the Exploration of the Sea).
}
\details{
\emph{Calculate PA reference points:}
\tabular{ll}{
  \code{\link{Bpa}} \tab from Blim\cr
  \code{\link{Fpa}} \tab from Flim
}
\emph{Calculate sigma:}
\tabular{ll}{
  \code{\link{sigmaCI}} \tab from confidence interval\cr
  \code{\link{sigmaPA}} \tab from PA reference points
}
}
\note{
\code{browseVignettes("icesAdvice")} shows the development guidelines for
contributing to this package.
}
\author{
Arni Magnusson, with contributions from Colin Millar.
}
\references{
Overview of the ICES advisory process:
\url{http://ices.dk/community/advisory-process/Pages/default.aspx}.

ICES list of software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.
}

