% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesAdvice-package.R
\docType{package}
\name{icesAdvice-package}
\alias{icesAdvice-package}
\alias{icesAdvice}
\title{Functions Related to ICES Advice}
\description{
Functions that are related to the ICES advisory process.
}
\details{
\emph{Evaluate ICES advice:}
\tabular{ll}{
  \code{\link{DLS3.2}}    \tab DLS method 3.2\cr
  \code{\link{icesRound}} \tab rounding method
}
\emph{Calculate PA reference points:}
\tabular{ll}{
  \code{\link{Bpa}} \tab from Blim\cr
  \code{\link{Fpa}} \tab from Flim
}
\emph{Calculate sigma:}
\tabular{ll}{
  \code{\link{sigmaCI}} \tab from confidence interval\cr
  \code{\link{sigmaPA}} \tab from PA reference points
}
}
\references{
ICES advisory process:
\url{http://ices.dk/community/advisory-process/}.
}
\author{
Arni Magnusson and Anne Cooper, with contributions by Colin Millar.
}
